make more platform agnostic

+++ src/core/offsets/Dumper.hpp
@@ -1,8 +1,8 @@
 #pragma once
-#include "core/memory/Memory.hpp"
+#include "core/memory/IMemory.hpp"
 #include "Offsets.hpp"
 
-inline const DWORD MAX_BLOCK_SIZE = 409600;
+inline const size_t MAX_BLOCK_SIZE = 409600;
 
 class Dumper {
 public:
@@ -25,8 +25,8 @@ private:
     bool InitImpl();
 private:
     std::vector<WORD> StrSigToArray(const std::string& sig);
-    DWORD64 Scan(const std::string sig, ProcessModule module);
+    uintptr_t Scan(const std::string sig, ProcessModule module);
     void GetNextArray(std::vector<short>& next, const std::vector<WORD>& signature);
-    std::vector<DWORD64> ScanMemory(const std::string& sig, DWORD64 start, DWORD64 end, int number = 1);
-    void ScanBlock(byte* buffer, const std::vector<short>& next, const std::vector<WORD>& signature, DWORD64 start, DWORD size, std::vector<DWORD64>& result);
+    std::vector<uintptr_t> ScanMemory(const std::string& sig, uintptr_t start, uintptr_t end, int number = 1);
+    void ScanBlock(byte* buffer, const std::vector<short>& next, const std::vector<WORD>& signature, uintptr_t start, size_t size, std::vector<uintptr_t>& result);
 };
